+++ cdlclient/__init__.py
@@ -13,10 +13,11 @@ application through XML-RPC protocol.
 .. _DataLab: https://codra-ingenierie-informatique.github.io/DataLab/
 """
 
+# pylint: disable=unused-import
 from cdlclient.baseproxy import SimpleBaseProxy
 from cdlclient.remote import SimpleRemoteProxy
 
-__version__ = "0.4.0"
+__version__ = "0.5.0"
 __docurl__ = "https://cdlclient.readthedocs.io/en/latest/"
 __homeurl__ = "https://github.com/Codra-Ingenierie-Informatique/DataLabSimpleClient/"
 __supporturl__ = "https://github.com/Codra-Ingenierie-Informatique/DataLabSimpleClient/issues/new/choose"