+++ cdlclient/tests/remoteclient_app.py
@@ -15,18 +15,17 @@ DataLab Remote client application test
 from __future__ import annotations
 
 import functools
-from collections.abc import Callable
 from contextlib import contextmanager
 
 import numpy as np
 from guidata.env import execenv
-from guidata.qthelpers import qt_app_context, qt_wait, win32_fix_title_bar_background
-from qtpy import QtCore as QC
+from guidata.qthelpers import qt_app_context, qt_wait
 from qtpy import QtWidgets as QW
 
 from cdlclient import SimpleRemoteProxy
 from cdlclient.tests.remoteclient_base import AbstractClientWindow
 from cdlclient.tests.remoteclient_unit import multiple_commands
+from cdlclient.widgets import DataLabConnectionDialog
 
 APP_NAME = "Remote client test"
 
@@ -56,80 +55,6 @@ def try_send_command():
     return try_send_command_decorator
 
 
-class DataLabConnectionThread(QC.QThread):
-    """DataLab Connection thread"""
-
-    SIG_CONNECTION_OK = QC.Signal()
-    SIG_CONNECTION_KO = QC.Signal()
-
-    def __init__(self, connect_callback: Callable, parent: QC.QObject = None) -> None:
-        super().__init__(parent)
-        self.connect_callback = connect_callback
-
-    def run(self) -> None:
-        """Run thread"""
-        try:
-            self.connect_callback()
-            self.SIG_CONNECTION_OK.emit()
-        except ConnectionRefusedError:
-            self.SIG_CONNECTION_KO.emit()
-
-
-class DataLabConnectionDialog(QW.QDialog):
-    """DataLab Connection dialog
-
-    Args:
-        connect_callback: Callback function to connect to DataLab server
-        parent: Parent widget. Defaults to None.
-    """
-
-    def __init__(self, connect_callback: Callable, parent: QW.QWidget = None) -> None:
-        super().__init__(parent)
-        win32_fix_title_bar_background(self)
-        self.setWindowTitle("Connection to DataLab")
-        self.setMinimumWidth(300)
-        self.host_label = QW.QLabel("Host:")
-        self.progress_bar = QW.QProgressBar()
-        self.progress_bar.setRange(0, 0)
-        self.status_label = QW.QLabel("Waiting for connection...")
-        layout = QW.QVBoxLayout()
-        layout.addWidget(self.host_label)
-        layout.addWidget(self.progress_bar)
-        layout.addWidget(self.status_label)
-        self.setLayout(layout)
-        self.thread = DataLabConnectionThread(connect_callback)
-        self.thread.SIG_CONNECTION_OK.connect(self.on_connection_successful)
-        self.thread.SIG_CONNECTION_KO.connect(self.on_connection_failed)
-        button_box = QW.QDialogButtonBox(QW.QDialogButtonBox.Cancel)
-        button_box.rejected.connect(self.reject)
-        layout.addWidget(button_box)
-
-    def exec(self) -> int:
-        """Execute dialog"""
-        self.connect_to_server()
-        return super().exec()
-
-    def connect_to_server(self) -> None:
-        """Connect to server"""
-        self.progress_bar.setRange(0, 0)
-        self.status_label.setText("Connecting to server...")
-        self.thread.start()
-
-    def on_connection_successful(self) -> None:
-        """Connection successful"""
-        self.progress_bar.setRange(0, 1)
-        self.progress_bar.setValue(1)
-        self.status_label.setText("Connection successful!")
-        self.accept()
-
-    def on_connection_failed(self) -> None:
-        """Connection failed"""
-        self.progress_bar.setRange(0, 1)
-        self.progress_bar.setValue(1)
-        self.status_label.setText("Connection failed.")
-        self.reject()
-
-
 class HostWindow(AbstractClientWindow):
     """Test main view"""
 
@@ -149,7 +74,6 @@ class HostWindow(AbstractClientWindow):
         if self.cdl is None:
             self.cdl: SimpleRemoteProxy = SimpleRemoteProxy()
             connect_dlg = DataLabConnectionDialog(self.cdl.connect, self)
-            connect_dlg.host_label.setText("Host: DataLab server")
             ok = connect_dlg.exec()
             if ok:
                 self.host.log("✨ Initialized DataLab connection ✨")