+++ CHANGELOG.md
@@ -1,10 +1,23 @@
 # DataLab Simple Client Releases #
 
+## Version 0.5.0 ##
+
+💥 Changes:
+
+* Remote API (`SimpleRemoteProxy`):
+  * Added `is_connected` method
+
+* New `widgets` module:
+  * New `DataLabConnectionDialog` class:
+    * Ready-to-use dialog box to connect to a DataLab server
+    * `from cdlclient.widgets import DataLabConnectionDialog`
+    * See example in `cdlclient/tests/connect_dialog.py`
+
 ## Version 0.4.0 ##
 
-Changes:
+💥 Changes:
 
-* Remote API:
+* Remote API (`SimpleRemoteProxy`):
   * Added dict-like interface (e.g. `proxy['obj123']`)
   * Renamed `switch_to_panel` method to `set_current_panel` (compatibility with DataLab 0.9)
   * Added `get_current_panel` method
@@ -13,9 +26,9 @@ Changes:
 
 ## Version 0.3.0 ##
 
-Changes:
+💥 Changes:
 
-* Remote API:
+* Remote API (`SimpleRemoteProxy`):
   * `get_object` method now takes either object number, UUID or a title
   * `get_object_shapes` method now takes either object number, UUID or a title
   * Removed deprecated `get_object_from_uuid` and `get_object_from_title` methods
@@ -26,9 +39,9 @@ Changes:
 
 ## Version 0.2.0 ##
 
-Changes:
+💥 Changes:
 
-* Remote API:
+* Remote API (`SimpleRemoteProxy`):
   * New `raise_window` method
   * New `get_object_shapes` method
   * New `get_object` method